requirements.txt
langchain==1.2.15
langgraph==1.1.6
google-generativeai
streamlit==1.56.0
python-dotenv==1.2.2
python-dateutil>=2.9.0.post0
reportlab
pathlib2>=2.3.7
groq
psycopg2-binary==2.9.10